orientation-event/DeviceOrientation/horizontal-surface.html

<!DOCTYPE html>
<html>

  <head>
    <title>02 - A device lying flat on a horizontal surface:</title>
    <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>02 - A device lying flat on a horizontal surface:</h1>
    <div id="log">
    </div>
	
	<script>
	
	function init(){
	   if (window.DeviceOrientationEvent) {
	  		window.addEventListener('deviceorientation', function(eventData) {
	
				// Gamma : angle par rapport a x
				gamma = eventData.gamma;
				// Beta : angle par rapport a y
				beta = eventData.beta;
				// Alpha : orientation (N-S-E-O)
				alpha = eventData.alpha;
				
			}, false);
			
		}
	}

	init();
	
	alert("Put your device on a horizontal surface.");
	
	test(function() {assert_approx_equals(gamma, 0, 2)}, "X angle");
	test(function() {assert_approx_equals(beta, 0, 2)}, "Y angle");
	  
	</script>
	
  </body>
  
</html>
